+++ apps/react/src/components/step.tsx
@@ -1,19 +1,23 @@
-import type { ReactElement } from "react";
-import type { UseFormProps } from "react-hook-form";
+import type { ReactNode } from "react";
+import type { DefaultValues, Resolver } from "react-hook-form";
 
 import { FormProvider, useForm } from "react-hook-form";
 
 import { useMultiStep } from "../multi-step";
 
-interface StepProps {
-  defaultValues: UseFormProps["defaultValues"];
-  resolver: UseFormProps["resolver"];
-  children: ReactElement;
+interface StepProps<T extends Record<string, unknown>> {
+  defaultValues: DefaultValues<T>;
+  resolver: Resolver<T>;
+  children: ReactNode;
 }
 
-export default function Step({ defaultValues, resolver, children }: StepProps) {
+export default function Step<T extends Record<string, unknown>>({
+  defaultValues,
+  resolver,
+  children,
+}: StepProps<T>) {
   const form = useForm({ defaultValues, resolver });
-  const { onNext } = useMultiStep();
+  const { onNext } = useMultiStep<T>();
   return (
     <form onSubmit={form.handleSubmit(onNext)} className="h-full">
       <FormProvider {...form}>{children}</FormProvider>